requirements.txt
requests
requests-oauthlib
python-dotenv
gspread
google-auth
google-auth-oauthlib
reportlab
google-api-python-client
